refactor: split code into packages
