+++ app/controller/nextdate.go
@@ -1,0 +1,33 @@
+package controller
+
+import (
+	"fmt"
+	"main/app/model"
+	"main/app/service"
+	"net/http"
+	"time"
+)
+
+func GetNextDate(w http.ResponseWriter, r *http.Request) {
+	now, err := time.Parse(model.DatePattern, r.FormValue("now"))
+	if err != nil {
+		http.Error(w, fmt.Sprintf(""), http.StatusBadRequest)
+		return
+	}
+
+	date := r.FormValue("date")
+	repeat := r.FormValue("repeat")
+	nextDate, err := service.NextDate(now, date, repeat)
+
+	if err != nil {
+		http.Error(w, fmt.Sprintf(""), http.StatusBadRequest)
+		return
+	}
+
+	w.WriteHeader(http.StatusOK)
+	_, err = w.Write([]byte(nextDate))
+
+	if err != nil {
+		http.Error(w, fmt.Errorf("writing tasks data error: %w", err).Error(), http.StatusBadRequest)
+	}
+}
