+++ app/model/error.go
@@ -1,0 +1,5 @@
+package model
+
+type ErrorResponse struct {
+	Error string `json:"error"`
+}
